requirements.txt
requests
typing
streamlit
python-dotenv==1.0.1
